moved alignPDBEnsemble to legacy
