+++ prody/utilities/legacy.py
@@ -610,3 +610,81 @@ def _extend(self, arr, defval=0):
     else: # only developers can trigger this case
         raise ValueError('arr can only be either 1D or 2D')
     return whole_array
+
+def alignPDBEnsemble(ensemble, suffix='_aligned', outdir='.', gzip=False):
+    """Align PDB files using transformations from *ensemble*, which may be
+    a :class:`.PDBEnsemble` or a :class:`.PDBConformation` instance. Label of
+    the conformation (see :meth:`~.PDBConformation.getLabel`) will be used to
+    determine the PDB structure and model number.  First four characters of
+    the label is expected to be the PDB identifier and ending numbers to be the
+    model number.  For example, the :class:`.Transformation` from conformation
+    with label *2k39_ca_selection_'resnum_<_71'_m116* will be applied to 116th
+    model of structure **2k39**.  After applicable transformations are made,
+    structure will be written into *outputdir* as :file:`2k39_aligned.pdb`.
+    If ``gzip=True``, output files will be compressed.  Return value is
+    the output filename or list of filenames, in the order files are processed.
+    Note that if multiple models from a file are aligned, that filename will
+    appear in the list multiple times."""
+
+    if not isinstance(ensemble, (PDBEnsemble, PDBConformation)):
+        raise TypeError('ensemble must be a PDBEnsemble or PDBConformation')
+    if isinstance(ensemble, PDBConformation):
+        ensemble = [ensemble]
+    if gzip:
+        gzip = '.gz'
+    else:
+        gzip = ''
+    output = []
+    pdbdict = {}
+    for conf in ensemble:
+        trans = conf.getTransformation()
+        if trans is None:
+            raise ValueError('transformations are not calculated, call '
+                             '`superpose` or `iterpose`')
+        label = conf.getLabel()
+
+        pdb = label[:4]
+        filename = pdbdict.get(pdb, fetchPDB(pdb))
+        if filename is None:
+            LOGGER.warning('PDB file for conformation {0} is not found.'
+                           .format(label))
+            output.append(None)
+            continue
+        LOGGER.info('Parsing PDB file {0} for conformation {1}.'
+                    .format(pdb, label))
+
+        acsi = None
+        model = label.rfind('m')
+        if model > 3:
+            model = label[model+1:]
+            if model.isdigit():
+                acsi = int(model) - 1
+            LOGGER.info('Applying transformation to model {0}.'
+                        .format(model))
+
+        if isinstance(filename, str):
+            ag = parsePDB(filename)
+        else:
+            ag = filename
+
+        if acsi is not None:
+            if acsi >= ag.numCoordsets():
+                LOGGER.warn('Model number {0} for {1} is out of range.'
+                            .format(model, pdb))
+                output.append(None)
+                continue
+            ag.setACSIndex(acsi)
+        trans.apply(ag)
+        outfn = os.path.join(outdir, pdb + suffix + '.pdb' + gzip)
+        if ag.numCoordsets() > 1:
+            pdbdict[pdb] = ag
+        else:
+            writePDB(outfn, ag)
+        output.append(os.path.normpath(outfn))
+
+    for pdb, ag in pdbdict.items():  # PY3K: OK
+        writePDB(os.path.join(outdir, pdb + suffix + '.pdb' + gzip), ag)
+    if len(output) == 1:
+        return output[0]
+    else:
+        return output
